+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 v1.6.6
+ - 感谢 @encircles 的pr,使用FuncWrapFieldTagName函数自定义Tag列名
  - 简化查询逻辑,提升性能
  - 修复获取自增主键异常
  - 完善文档,注释